Forbildungsfront + backend, submit form in bearbeitung

+++ src/main/java/org/fobidb/fortbildungen/FortbildungRepository.java
@@ -8,9 +8,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface FortbildungRepository extends JpaRepository<Fortbildung, Integer> { // unser TeacherRepository erbt von JpaRepository, daher stehen viele Befehle zum manipulieren der Datenbank zur Verfügung
+public interface FortbildungRepository extends JpaRepository<Fortbildung, Long> { // unser TeacherRepository erbt von JpaRepository, daher stehen viele Befehle zum manipulieren der Datenbank zur Verfügung
 
     //SELECT * FROM teacher WHERE email = ?
     @Query("SELECT f FROM Fortbildung f WHERE f.Fortbildung_ID = ?1")
-    Optional<Fortbildung> findByFobid_ID(int id);
+    Optional<Fortbildung> findByFobid_ID(Long id);
 }
